Separate the components in different files
